Run the `serve` module as a service on Alpine.
Run the gather and render modules on Alpine every five minutes.  Add the
`update_trmnl` script to run the modules and add the `trmnl` crontab
that runs the update script on schedule.

Run the serve module on Alpine as a service.  Add the `serve_trmnl`
script to run the module and add the `trmnl_srv` service script.

Update the serve module's `Server` class to allow address and port
reuse.  Update the makefile to deploy the new Alpine files.