Render bitmap using Atkinson Hyperlegible Mono.
Include the Atkinson Hyperlegible Mono OpenType font in the `render`
package and load it when rendering.

Add the `Screen` class to `render`.  The `Screen` class calculates key
screen dimensions, dividing the 800x480 pixel bitmap up into cells for
rendering monospaced characters, and it also provides the 2D grid of
character cells.

Added `write()` and `write_reverse()` methods for writing strings to the
grid.  The `write()` method starts at the given column index and writes
characters to the right; the `write_reverse()` method writes characters
to the left.

Convert the `updated` time from the JSON content file to Pacific time
and format it more nicely.

Split the weather forecast details up into lines and render them as a
paragraph.