Rework to use nginx to serve mostly static data.
Change the `render` module to also write out the JSON response returned
by the /api/display endpoint.  Set the `filename` key to the updated
time, since the TRMNL looks at this value to determine if the bitmap
image has changed and should be fetched.

Also change `render` to generate a PNG as well as a BMP.  Testing showed
that TRMNL happily reads PNGs.

Currently, `render` still writes all its output to a temp directory, so
add the `update` script to copy the `render` output files from the temp
location to the appropriate place in the new `static/` directory, which
is the root of files now served by nginx.

Tested nginx locally on my mac; the nginx config is included in the
`docs/` directory.

Added `static/` JSON files `api/display/index.json` (which is
regenerated by `render` and `api/setup/index.json` (which doesn't need
to change, but _does_ contain the URL of the `trmnl_srv` and will need
to be generated once for a given server.  Also added `bitmap/index.png`
default bitmap file.  The nginx config will serve up `index.json` and
`index.png` as default index pages.

Changed the `serve` module to listen on port 4001, where it is proxied
by nginx on port 4000.