Move web content to `container/srv/www/`.  Add some CSS.
Move the files from `static/` to `container/srv/www/`.  Add a top-level
index HTML page with links to `/api/setup`, `/api/display` and `/bitmap`
for inspection.  Add CSS and fonts for the HTML page.

Generate a "welcome" bitmap and tweak the JSON for the default web
content.  Add a `write_intro_screen()` function to the `render` package
to generate the welcome bitmap.