Run Python apps in the container.
Add missing timezone package to the container.  Copy cron jobs and
Python code and set the entrypoint to the `trmnl_srv` script, which runs
`serve`, `nginx` and `crond`.

Copy the default Alpine `root` crontab and add a job to run `gather` and
`render`.

Clean up some old scripts that are no longer needed.