Add `--base-url` option to `render`.
When writing the `/api/display/index.json` and `/api/setup/index.json`
files, dynamically set the base URL for the `image_url` keys.

Modify the `gather-render` control script to read the `--base-url`
option from the environment.

In the Dockerfile, add `ARG BASE_URL` and rewrite the placeholder
`index.json` files to use this value.

Pass the `BASE_URL` value to `docker build` as a build arg and to
`docker run` as an env value.  Also make the `docker run` publish port a
makefile variable.