Grab hourly forecast for evaluation.
Add a `~source_data` key to the content JSON, tilde prefixed so it sorts
last.  Move the current `forecast_response` and `points_response` keys
from `Weather` there and add the `forecast_hourly_response` key.

Also fix a bug in `gather` that was saving the `points_response` in
place of the `forecast_response`.