Add `LogStorage` class.
Make the `Options` classes regular classes instead of dataclasses and
add initializers; create a `LogStorage` attribute named `trmnl_logs` in
`CommonOptions` initialized with the `web_root` parameter.

In `gather`, pass the `trmnl_logs` value to `get_log()` in place of
`web_root` and remove the `is_log_file()` and `list_log_files()` helper
functions, moving this functionality to `LogStorage`.

Also add a `trmnl_logs` attribute to `Server` in `serve` and use it in
the `Handler` when writing new log messages.

Rename the `LogEntry` class to `TRMNLLogEntry` and move it to the
`common.log` package.